// src/app/models/OperacionPrincipal.model.ts
export class OperacionPrincipalModel {
  Id_OperacionPrincipal: any;
  DescripcionOperacion: string;
  FechaOperacion: Date;
  ResponsableGiro: string;
  CodigoOperacion: string;
  Numero: number;
  Ano: number;
  Mes: string;
  Id_Proyecto: number;
  C01: number;
  C02: number;
  C03: number;
  C04: number;
  C06: number;
  C07: number;
  C08: number;
  C09: number;
  C10: number;
  C11: number;
  C12: number;
  C13: number;
  C14: number;
  C16: number;
  C17: number;
  C18: number;
  C19: number;
  C20: number;
  C21: number;
  C22: number;
  C23: number;
  C24: number;
  C25: number;
  C26: number;
  C27: number;
  C28: number;
  C29: number;
  C30: number;
  C31: number;
  C32: number;
  C33: number;
  C34: number;
  C35: number;
  C36: number;
  C37: number;
  C38: number;
  C39: number;
  C40: number;
  C41: number;
  C42: number;
  C43: number;
  C44: number;
  C45: number;
  C46: number;
  C47: number;
  C48: number;
  C49: number;
  C50: number;
  C51: number;
  C52: number;
  C56: number;
  C57: number;
  C58: number;
  C59: number;
  C60: number;
  C61: number;
  C62: number;
  C63: number;
  C64: number;
  C65: number;
  C66: number;
  C67: number;
  C68: number;
  C69: number;
  C70: number;
  C71: number;
  C72: number;
  C73: number;
  C74: number;
  C75: number;
  C76: number;
  C77: number;
  C78: number;
  C79: number;
  C80: number;
  C81: number;
  C82: number;
  C83: number;
  C84: number;
  C85: number;
  C87: number;
  C88: number;
  C89: number;
  C91: number;
  C92: number;
  C93: number;
  C94: number;
  C95: number;
  C96: number;
  TipoOrigen: string;
  Solicitud: number;
  AutorizacionDolares: number;
  AutorizacionSoles: number;
  Diferencia: number;
  Cuadrado: Boolean;
  Cerrado: boolean;

  constructor() {
    this.Id_OperacionPrincipal = null;
    this.DescripcionOperacion = '';
    this.FechaOperacion = new Date();
    this.ResponsableGiro = '';
    this.CodigoOperacion = '';
    this.Numero = 0;
    this.Ano = 0;
    this.Mes = '';
    this.Id_Proyecto = 0;
    this.C01 = 0;
    this.C02 = 0;
    this.C03 = 0;
    this.C04 = 0;
    this.C06 = 0;
    this.C07 = 0;
    this.C08 = 0;
    this.C09 = 0;
    this.C10 = 0;
    this.C11 = 0;
    this.C12 = 0;
    this.C13 = 0;
    this.C14 = 0;
    this.C16 = 0;
    this.C17 = 0;
    this.C18 = 0;
    this.C19 = 0;
    this.C20 = 0;
    this.C21 = 0;
    this.C22 = 0;
    this.C23 = 0;
    this.C24 = 0;
    this.C25 = 0;
    this.C26 = 0;
    this.C27 = 0;
    this.C28 = 0;
    this.C29 = 0;
    this.C30 = 0;
    this.C31 = 0;
    this.C32 = 0;
    this.C33 = 0;
    this.C34 = 0;
    this.C35 = 0;
    this.C36 = 0;
    this.C37 = 0;
    this.C38 = 0;
    this.C39 = 0;
    this.C40 = 0;
    this.C41 = 0;
    this.C42 = 0;
    this.C43 = 0;
    this.C44 = 0;
    this.C45 = 0;
    this.C46 = 0;
    this.C47 = 0;
    this.C48 = 0;
    this.C49 = 0;
    this.C50 = 0;
    this.C51 = 0;
    this.C52 = 0;
    this.C56 = 0;
    this.C57 = 0;
    this.C58 = 0;
    this.C59 = 0;
    this.C60 = 0;
    this.C61 = 0;
    this.C62 = 0;
    this.C63 = 0;
    this.C64 = 0;
    this.C65 = 0;
    this.C66 = 0;
    this.C67 = 0;
    this.C68 = 0;
    this.C69 = 0;
    this.C70 = 0;
    this.C71 = 0;
    this.C72 = 0;
    this.C73 = 0;
    this.C74 = 0;
    this.C75 = 0;
    this.C76 = 0;
    this.C77 = 0;
    this.C78 = 0;
    this.C79 = 0;
    this.C80 = 0;
    this.C81 = 0;
    this.C82 = 0;
    this.C83 = 0;
    this.C84 = 0;
    this.C85 = 0;
    this.C87 = 0;
    this.C88 = 0;
    this.C89 = 0;
    this.C91 = 0;
    this.C92 = 0;
    this.C93 = 0;
    this.C94 = 0;
    this.C95 = 0;
    this.C96 = 0;
    this.TipoOrigen = '';
    this.Solicitud = 0;
    this.AutorizacionDolares = 0;
    this.AutorizacionSoles = 0;
    this.Diferencia = 0;
    this.Cuadrado = false;
    this.Cerrado = false;
  }
}